Correção -> Sistema não deixava abrir tela de realização de importação de backup mais de uma vez. Issue #3 Correção -> Atualizar versão do sistema fazia necessário a escolha do banco de dados. Agora a versão é salva em registro único.
Melhoria -> Sistema deixa visualizar o arquivo de log através do menu Ajuda! Issue #4
Melhoria -> Código inspecionado automaticamente para correções.

+++ src/snu/util/ProcessadorDeConsultasUtil.java
@@ -63,10 +63,10 @@ public class ProcessadorDeConsultasUtil {
             }
         }
 
-        TF = (double) tiq / (double) tmax;
+        TF = tiq / tmax;
 
         IDF = Math.log(quantidadeDocumentos.doubleValue()
-                / ((double) vocabulo.getListaInvertida().size()))
+                / vocabulo.getListaInvertida().size())
                 / Math.log(2.0);
 
         peso = TF * IDF;
